Cloud. Websocket. Communication.

+++ config_flow.py
@@ -40,7 +40,7 @@ from homeassistant.data_entry_flow import FlowResult
 user_step_data_schema = {
     vol.Required(CONF_USERNAME, default="frederick.stallmeyer1@qconnex.com"): str,
     vol.Required(CONF_PASSWORD): str,
-    vol.Required(CONF_SCAN_INTERVAL, default=60): float,
+    vol.Required(CONF_SCAN_INTERVAL, default=60): int,
     vol.Required(CONF_SYNC_ROOMS, default=True): bool,
     vol.Required(CONF_HOST, default="https://app-api.test.qconnex.io"): str,
 }
